selling proccess Payment procecc with/without debts Debitors list with pagination and some other endpoints added

+++ src/main/java/shakh/billingsystem/models/OrderItemsDto.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public class OrderItemsDto {
     private Long productId;
     private String productName;
-    private Integer barcode;
+    private Long barcode;
     private Double amount;
     private Double priceOfSell;
     private Double priceOfBuy;
